Fix gravity block position at high coordinates
See https://github.com/retromcorg/Project-Poseidon/commit/02df2eb8e0844e086fc0ff286da8d8cb626028e9

+++ poseidon-server/src/main/java/net/minecraft/server/BlockSand.java
@@ -27,7 +27,8 @@ public class BlockSand extends Block {
             byte b0 = 32;
 
             if (!instaFall && world.a(i - b0, j - b0, k - b0, i + b0, j + b0, k + b0)) {
-                EntityFallingSand entityfallingsand = new EntityFallingSand(world, (float) i + 0.5F, (float) j + 0.5F, (float) k + 0.5F, this.id);
+                // Poseidon - fix gravity block position at high coordinates
+                EntityFallingSand entityfallingsand = new EntityFallingSand(world, i + 0.5D, j + 0.5D, k + 0.5D, this.id);
 
                 world.addEntity(entityfallingsand);
             } else {
